fix(supreme): add missing .toInt() on 0x94000000 Long literal
Last remaining build error:
  UnlimitedCoinsPatch.kt:136 'Operator == cannot be applied to Int and Long'

0x94000000 (BL opcode) exceeds Int.MAX_VALUE so Kotlin infers Long.
Added .toInt() to match the other hex literals.

+++ patches/src/main/kotlin/com/Neurononfire/SupremeDuelist/patches/iap/UnlimitedCoinsPatch.kt
@@ -133,7 +133,7 @@ val unlimitedCoinsPatch = rawResourcePatch(
                     // If we hit a branch/ret before finding the str, give up.
                     // Use .toInt() to convert Long literals (> Int.MAX_VALUE) to Int.
                     val isB = (word and 0xFC000000.toInt()) == 0x14000000
-                    val isBl = (word and 0xFC000000.toInt()) == 0x94000000
+                    val isBl = (word and 0xFC000000.toInt()) == 0x94000000.toInt()
                     val isRet = (word and 0xFFFFFC00.toInt()) == 0xD65F0000.toInt()
                     if (isB || isBl || isRet) {
                         break
